Moved override tests
